requirements.txt
Django==5.2.7
requests==2.32.5
python-decouple==3.8
PyPDF2==3.0.1
python-docx==1.2.0
pillow==12.0.0
